fix: misc improvements to tsconfig
remove extraneous and irrelevant 'lib' options. sets `inlineSourcemap`
to false, as sourcemaps created externally can still be handled fine. add
`useDefineForClassFields` and set to to ensure class field and getter
handling are more compliant with the ECMA spec (at a later release of
typescript this will be defaulted to true, so this will decrease
potential for bugs to crop up during an upgrade)